+++ app/src/main/java/com/pengxh/autodingding/utils/OnTimeCountDownCallback.kt
@@ -1,7 +1,7 @@
 package com.pengxh.autodingding.utils
 
 interface OnTimeCountDownCallback {
-    fun updateCountDownSeconds(remainingSeconds: Long)
+    fun updateCountDownSeconds(remainingSeconds: Int)
 
     fun onFinish()
 }
